fix connection broken detection (on ssl sockets)
The previous checks within `Connection::readFrame` often failed for ssl
connections on AWS. This was caused by the fact that `stream_select` will
fail to reliable determine if there is already processable data on the
encrypted stream. It returns true while the stream is not yet ready to
be consumed, which causes `fread` to return an empty string, which was
considered as failure before.
Related to https://github.com/stomp-php/stomp-php/issues/115

BC: As `fread` also returns an empty string for sockets that are broken,
we now need to make use of heartbeats to determine the status of the
connection. `HeartbeatEmitter` was adapted, so that a failing alive
notification now also causes an exception.
In addition `ServerAliveObserver` was added, which simplifies the usage
of server side triggered heartbeats to determine if the connection is
still stable.