Fix activemq.subscriptionName
To create a durable subscription the client-id header must be set on the CONNECT frame and the durable-subscription-name must be set on the SUBSCRIBE frame. The combination of these two headers will form the identity of the durable subscription.

Source https://activemq.apache.org/components/artemis/documentation/2.7.0/stomp.html#durable-subscriptions

So instead of sending the client-id in the SUBSCRIBE frame, we should send the subscription id.